RAG Notebook
Just RAG portion, this notebook defines the Bartenders "behavior" in the database.
It's current characteristics allow it to be pleasant, modest, empathetic,  a little "self-aware", and understanding by recommending drinks or commenting on a customers order all through the retrieval of data in the database and providing and augmented generation if relevant. 
You can add more behaviors by simply supplying a quote into the data set.
Notebook also includes a test section, if you'd like to test it, simply update the query portion.